Revert removal of translated attributes from default tables
We are migrating from Globalize to Mobility.

Mobility does not raise a warning in every query for having the original attributes in their default tables, apart from having them in the corresponding translation table. Which is great.

Developers expect to see attributes in their default tables, regardless of any
translation funcionality. Potential conflicts between the values in the default and translated table should be dealt with at a code level not at a database level.

In this commit we are bringing back the text attributes of every translated table to the corresponding default tables.